+++ docs/source/examples/graph_based/plot_1_dl2vec.py
@@ -27,88 +27,29 @@ sys.path.append('../../')
 import mowl
 mowl.init_jvm("10g")
 
-from mowl.datasets.builtin import GDAMouseDataset
+from mowl.datasets.builtin import GDADatasetV2
+from mowl.models import RandomWalkPlusW2VModel
 from mowl.projection import DL2VecProjector
 from mowl.walking import DeepWalk
 from gensim.models.word2vec import LineSentence
 from gensim.models import Word2Vec
 
-
 # %%
-# Projecting the ontology
-# -----------------------
-#
-# We project the ontology using the DL2VecProjector class. The rules used to project the 
-# ontology can be found at :doc:`/graphs/projection`. The outcome of the projection algorithm
-# is an edgelist.
+# Instantiating the dataset and the model
 
-dataset = GDAMouseDataset()
+dataset = GDADatasetV2()
 
-projector = DL2VecProjector(bidirectional_taxonomy=True)
-edges = projector.project(dataset.ontology)
+print(f"Number of classes: f{len(dataset.classes)}")
 
+model = RandomWalkPlusW2VModel(dataset)
+model.set_projector(DL2VecProjector())
+model.set_walker(DeepWalk(5, 5, 0.1, workers=4))
+model.set_w2v_model(vector_size=5, epochs=2, window=5, min_count=1, workers=4)
+model.train()
 # %%
-# Generating random walks
-# -----------------------
-#
-# The random walks are generated using the DeepWalk class. This class implements the DeepWalk
-# algorithm with a modification consisting of including the edge labels as part of the walks.
+# Evaluating the model
 
-walker = DeepWalk(20, # number of walks per node
-                  20, # walk length
-                  0.1, # restart probability
-                  workers=4) # number of threads
+from mowl.evaluation import GDAEvaluator
+model.set_evaluator(GDAEvaluator)
+model.evaluate()
 
-walks = walker.walk(edges)
-
-
-# %%
-# Training the Word2Vec model
-# ---------------------------
-#
-# To train the Word2Vec model, we rely on the Gensim library:
-
-walks_file = walker.outfile
-sentences = LineSentence(walks_file)
-model = Word2Vec(sentences, vector_size=100, epochs = 20, window=5, min_count=1, workers=4)
-
-# %%
-# Evaluating the embeddings
-# ------------------------------
-#
-# We can evaluate the embeddings using the
-# :class:`EmbeddingsRankBasedEvaluator <mowl.evaluation.rank_based.EmbeddingsRankBasedEvaluator>`
-# class. We need to do some data preparation.
-
-from mowl.evaluation.rank_based import EmbeddingsRankBasedEvaluator
-from mowl.evaluation.base import CosineSimilarity
-from mowl.projection import TaxonomyWithRelationsProjector
-# %%
-# We are going to evaluate the plausability of an association gene-disease with a gene against all
-# possible diseases and check the rank of the true disease association.
-
-genes, diseases = dataset.evaluation_classes
-
-projector = TaxonomyWithRelationsProjector(taxonomy=False,
-                                           relations=["http://is_associated_with"])
-
-evaluation_edges = projector.project(dataset.testing)
-filtering_edges = projector.project(dataset.ontology)
-assert len(evaluation_edges) > 0
-
-# %%
-# The gene-disease associations will be scoredc using cosine similarity. For that reason we use
-# the ``CosineSimilarity`` class.
-
-vectors = model.wv
-evaluator = EmbeddingsRankBasedEvaluator(
-    vectors,
-    evaluation_edges,
-    CosineSimilarity,
-    training_set=filtering_edges,
-    head_entities = genes.as_str,
-    tail_entities = diseases.as_str,
-    device = 'cpu'
-)
-
-evaluator.evaluate(show=True)