+++ mowl/models/elembeddings/examples/model_ppi.py
@@ -17,6 +17,12 @@ class ELEmPPI(ELEmbeddings):
 
         self.set_evaluator(PPIEvaluator)
 
+    @property
+    def evaluation_model(self):
+        if self._evaluation_model is None:
+            self._evaluation_model = self.module
+
+        return self._evaluation_model
         
     def train(self, validate_every=1000):
 
@@ -78,4 +84,4 @@ class ELEmPPI(ELEmbeddings):
             metrics = self.evaluate()
             print(metrics)
                                     
-            
+